ADD [Drill] Animations

+++ src/sound/Sound.cpp
@@ -7,7 +7,7 @@ void Sound::play(double volume) {
 
   ma_sound_init_copy(ma_sound_get_engine(&_ma_sound), &_ma_sound, 0, nullptr, copy);
   ma_sound_start(copy);
-  ma_sound_set_volume(copy, volume);
+  ma_sound_set_volume(copy, static_cast<float>(volume));
 
   ma_sound_set_end_callback(
     copy,
